server/requirements.txt
fastapi
uvicorn
openai
anthropic
google-generativeai
groq
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
python-docx
reportlab
python-dotenv
httpx
beautifulsoup4
pydantic
duckduckgo-search
